Updates and file cleaning
