--- cycleGAN/train_test.py
@@ -1,391 +1,0 @@
-import torch
-import config
-import torch.nn as nn
-import torch.optim as optim
-from HE_IHC_dataset import GanDataset
-from torch.utils.data import DataLoader
-from tqdm import tqdm
-from discriminator import Discriminator
-from generator import Generator
-from evaluate import evaluate_fid_scores
-from utils import load_checkpoint, save_checkpoint, set_seed
-from torchvision.utils import save_image
-from torch.utils.tensorboard import SummaryWriter
-from datetime import datetime
-
-
-def train_func(D_HE, D_IHC, G_HE, G_IHC, optim_D, optim_G, G_scaler, D_scaler, cycle_loss, disc_loss, ident_loss, loader, epoch, writer):
-    
-    '''Set the generators and discriminators to training mode'''
-    D_HE.train()
-    D_IHC.train()
-    G_HE.train()
-    G_IHC.train()
-
-    loop = tqdm(loader, leave=True)         #Loop generates a progress bar while iterating over dataset
-    for idx, sample in enumerate(loop):
-        ihc = sample['A'].to(config.DEVICE)
-        he = sample['B'].to(config.DEVICE)
-
-        # ## DEBUGGING
-        # # Log the paths and indices
-        # A_path = sample["A_path"]
-        # B_path = sample["B_path"]
-        # A_index = sample['A_index']
-        # B_index = sample['B_index']
-        # patch_index = sample['patch_index']
-        # print(f"Epoch {epoch}, Batch {idx}: A_path: {A_path}, B_path: {B_path}, A_index: {A_index}, B_index: {B_index}, Patch_index: {patch_index}")
-
-        with torch.cuda.amp.autocast():     # For mixed precision training
-            '''Train the Discriminator of HE images'''
-            fake_HE = G_HE(ihc)
-            D_real_HE = D_HE(he)
-            D_fake_HE = D_HE(fake_HE.detach())
-
-            label_real_HE = torch.ones_like(D_real_HE).to(config.DEVICE)
-            label_fake_HE = torch.zeros_like(D_fake_HE).to(config.DEVICE)
-
-            D_HE_real_loss = disc_loss(D_real_HE, label_real_HE)
-            D_HE_fake_loss = disc_loss(D_fake_HE, label_fake_HE)
-            D_HE_loss = D_HE_real_loss + D_HE_fake_loss
-
-            '''Train the Discriminator of IHC images'''
-            fake_IHC = G_IHC(he)
-            D_real_IHC = D_IHC(ihc)
-            D_fake_IHC = D_IHC(fake_IHC.detach())
-
-            label_real_IHC = torch.ones_like(D_real_IHC).to(config.DEVICE)
-            label_fake_IHC = torch.zeros_like(D_fake_IHC).to(config.DEVICE)
-
-            D_IHC_real_loss = disc_loss(D_real_IHC, label_real_IHC)
-            D_IHC_fake_loss = disc_loss(D_fake_IHC, label_fake_IHC)
-            D_IHC_loss = D_IHC_real_loss + D_IHC_fake_loss
-
-            D_loss = (D_HE_loss + D_IHC_loss) / 2   # Using simple averaging for the discriminator loss
-
-            writer.add_scalar("[TRAIN] - HE Discriminator Loss", D_HE_loss, epoch)
-            writer.add_scalar("[TRAIN] - IHC Discriminator Loss", D_IHC_loss, epoch)
-            writer.add_scalar("[TRAIN] - Total Discriminator Loss", D_loss, epoch)
-
-        optim_D.zero_grad()
-        D_scaler.scale(D_loss).backward()
-        D_scaler.step(optim_D)
-        D_scaler.update()
-
-        with torch.cuda.amp.autocast():         # For mixed precision training
-            '''Train the Generator of HE and IHC images'''
-            D_fake_HE = D_HE(fake_HE)
-            D_fake_IHC = D_IHC(fake_IHC)
-
-            # Now the label for the generated images must be one (real) to fool the Discriminator
-            label_fake_HE = torch.ones_like(D_fake_HE).to(config.DEVICE)    
-            label_fake_IHC = torch.ones_like(D_fake_IHC).to(config.DEVICE)
-
-            D_gen_HE_loss = disc_loss(D_fake_HE, label_fake_HE)
-            D_gen_IHC_loss = disc_loss(D_fake_IHC, label_fake_IHC)
-
-            # Cycle Consistency Loss
-            cycle_IHC = G_IHC(fake_HE)
-            cycle_HE = G_HE(fake_IHC)
-            cycle_IHC_loss = cycle_loss(ihc, cycle_IHC)
-            cycle_HE_loss = cycle_loss(he, cycle_HE)
-
-            # Identity loss (remove these for efficiency if you set lambda_identity=0)
-            identity_IHC = G_IHC(ihc)
-            identity_HE = G_HE(he)
-            identity_IHC_loss = ident_loss(ihc, identity_IHC)
-            identity_HE_loss = ident_loss(he, identity_HE)
-
-            # Total generator loss
-            G_loss = (
-                D_gen_HE_loss
-                + D_gen_IHC_loss
-                + cycle_IHC_loss * config.LAMBDA_CYCLE
-                + cycle_HE_loss * config.LAMBDA_CYCLE
-                + identity_HE_loss * config.LAMBDA_IDENTITY
-                + identity_IHC_loss * config.LAMBDA_IDENTITY
-                )
-
-            writer.add_scalar("[TRAIN] - Cycle IHC Loss", cycle_IHC_loss, epoch)
-            writer.add_scalar("[TRAIN] - Cycle HE Loss", cycle_HE_loss, epoch)
-            writer.add_scalar("[TRAIN] - Identity IHC Loss", identity_IHC_loss, epoch)
-            writer.add_scalar("[TRAIN] - Identity HE Loss", identity_HE_loss, epoch)
-            writer.add_scalar("[TRAIN] - Fake_IHC Discriminator Loss", D_gen_IHC_loss, epoch)
-            writer.add_scalar("[TRAIN] - Fake_HE Discriminator Loss", D_gen_HE_loss, epoch)
-            writer.add_scalar("[TRAIN] - Total Generator Loss", G_loss, epoch)
-
-        optim_G.zero_grad()
-        G_scaler.scale(G_loss).backward()
-        G_scaler.step(optim_G)
-        G_scaler.update()
-
-        if epoch % 5 == 0 and idx % 1000 == 0:
-                for i in range(len(ihc)):   # (*0.5 + 0.5) before saving img to be on range [0, 1]
-                    save_image(he[i]*0.5 + 0.5, config.parent_path / f"gan-img/HE/train/epoch[{epoch}]_batch[{idx}]_HE[{i}].png")
-                    save_image(fake_HE[i]*0.5 + 0.5, config.parent_path / f"gan-img/HE/train/epoch[{epoch}]_batch[{idx}]_HE[{i}]_fake.png")
-                    save_image(ihc[i]*0.5 + 0.5, config.parent_path / f"gan-img/IHC/train/epoch[{epoch}]_batch[{idx}]_IHC[{i}].png")
-                    save_image(fake_IHC[i]*0.5 + 0.5, config.parent_path / f"gan-img/IHC/train/epoch[{epoch}]_batch[{idx}]_IHC[{i}]_fake.png")
-        
-        loop.set_postfix(D_loss=D_loss.item(), G_loss=G_loss.item())
-
-    print(f"\nTRAIN EPOCH: {epoch}/{config.NUM_EPOCHS}, batch: {idx+1}/{len(loader)}," + f" G_loss: {G_loss}, D_loss: {D_loss}\n")
-    
-    return G_loss, D_loss
-
-def eval_single_epoch(D_HE, D_IHC, G_HE, G_IHC, cycle_loss, disc_loss, ident_loss, loader, epoch, writer):
-
-    G_HE.eval()
-    G_IHC.eval()
-
-    loop = tqdm(loader, leave=True)
-    for idx, sample in enumerate(loop):
-        ihc = sample['A'].to(config.DEVICE)
-        he = sample['B'].to(config.DEVICE)
-
-        # ## DEBUGGING
-        # # Log the paths and indices
-        # A_path = sample['A_path']
-        # B_path = sample['B_path']
-        # A_index = sample['A_index']
-        # B_index = sample['B_index']
-        # patch_index = sample['patch_index']
-        # print(f"[VALIDATION] Epoch {epoch}, Batch {idx}: A_path: {A_path}, B_path: {B_path}, A_index: {A_index}, B_index: {B_index}, Patch_index: {patch_index}")
-
-        with torch.no_grad():
-            with torch.cuda.amp.autocast():   
-                fake_HE = G_HE(ihc)
-                fake_IHC = G_IHC(he)
-
-                D_fake_HE = D_HE(fake_HE)
-                D_fake_IHC = D_IHC(fake_IHC)
-
-                # Now the label for the generated images must be one (real) to fool the Discriminator
-                label_fake_HE = torch.ones_like(D_fake_HE).to(config.DEVICE)    
-                label_fake_IHC = torch.ones_like(D_fake_IHC).to(config.DEVICE)
-
-                D_gen_HE_loss = disc_loss(D_fake_HE, label_fake_HE)
-                D_gen_IHC_loss = disc_loss(D_fake_IHC, label_fake_IHC)
-
-                # Cycle Consistency Loss
-                cycle_IHC = G_IHC(fake_HE)
-                cycle_HE = G_HE(fake_IHC)
-                cycle_IHC_loss = cycle_loss(ihc, cycle_IHC)
-                cycle_HE_loss = cycle_loss(he, cycle_HE)
-
-                # Identity loss (remove these for efficiency if you set lambda_identity=0)
-                identity_IHC = G_IHC(ihc)
-                identity_HE = G_HE(he)
-                identity_IHC_loss = ident_loss(ihc, identity_IHC)
-                identity_HE_loss = ident_loss(he, identity_HE)
-
-                # Total generator loss
-                G_loss = (
-                    D_gen_HE_loss
-                    + D_gen_IHC_loss
-                    + cycle_IHC_loss * config.LAMBDA_CYCLE
-                    + cycle_HE_loss * config.LAMBDA_CYCLE
-                    + identity_HE_loss * config.LAMBDA_IDENTITY
-                    + identity_IHC_loss * config.LAMBDA_IDENTITY
-                )
-
-                writer.add_scalar("[VAL] - Cycle IHC Loss", cycle_IHC_loss, epoch)
-                writer.add_scalar("[VAL] - Cycle HE Loss", cycle_HE_loss, epoch)
-                writer.add_scalar("[VAL] - Identity IHC Loss", identity_IHC_loss, epoch)
-                writer.add_scalar("[VAL] - Identity HE Loss", identity_HE_loss, epoch)
-                writer.add_scalar("[VAL] - Fake_IHC Discriminator Loss", D_gen_IHC_loss, epoch)
-                writer.add_scalar("[VAL] - Fake_HE Discriminator Loss", D_gen_HE_loss, epoch)
-                writer.add_scalar("[VAL] - Total Generator Loss", G_loss, epoch)
-        
-        if epoch % 5 == 0 and idx % 170 == 0:
-                for i in range(len(ihc)):   # (*0.5 + 0.5) before saving img to be on range [0, 1]
-                    save_image(he[i]*0.5 + 0.5, config.parent_path / f"gan-img/HE/val/epoch[{epoch}]_batch[{idx}]_HE[{i}].png")
-                    save_image(fake_HE[i]*0.5 + 0.5, config.parent_path / f"gan-img/HE/val/epoch[{epoch}]_batch[{idx}]_HE[{i}]_fake.png")
-                    save_image(ihc[i]*0.5 + 0.5, config.parent_path / f"gan-img/IHC/val/epoch[{epoch}]_batch[{idx}]_IHC[{i}].png")
-                    save_image(fake_IHC[i]*0.5 + 0.5, config.parent_path / f"gan-img/IHC/val/epoch[{epoch}]_batch[{idx}]_IHC[{i}]_fake.png")
-
-    print(f"\nVALIDATION EPOCH: {epoch}/{config.NUM_EPOCHS}, batch: {idx+1}/{len(loader)}," + f" G_loss: {G_loss}\n")
-
-    return G_loss
-
-def custom_collate(batch):
-    # Initialize dictionaries to store batches for each key
-    batch_dict = {key: [] for key in batch[0]}
-    
-    # Append each item to the corresponding list in the batch_dict
-    for item in batch:
-        for key in item:
-            batch_dict[key].append(item[key])
-    
-    # Convert lists to tensors where applicable (A and B are tensors, others can be left as lists)
-    batch_dict['A'] = torch.stack(batch_dict['A'])
-    batch_dict['B'] = torch.stack(batch_dict['B'])
-    
-    return batch_dict
-
-def main():
-    set_seed(42) # To ensure reproducibility
-
-    disc_HE = Discriminator(in_channels=config.IN_CH, features=config.D_FEATURES).to(config.DEVICE) 
-    disc_IHC = Discriminator(in_channels=config.IN_CH, features=config.D_FEATURES).to(config.DEVICE)
-
-    gen_HE = Generator(img_channels=3, num_residuals=config.N_RES_BLOCKS).to(config.DEVICE)
-    gen_IHC = Generator(img_channels=3, num_residuals=config.N_RES_BLOCKS).to(config.DEVICE)
-
-    optim_disc = optim.Adam(
-        list(disc_HE.parameters()) + list(disc_IHC.parameters()),
-        lr=config.LEARNING_RATE,
-        betas=(0.5, 0.999)
-    )
-
-    optim_gen = optim.Adam(
-        list(gen_HE.parameters()) + list(gen_IHC.parameters()),
-        lr=config.LEARNING_RATE,
-        betas=(0.5, 0.999)
-    )
-
-    # Losses used during training
-    cycle_loss = nn.L1Loss()
-    identity_loss = nn.L1Loss()
-    discrim_loss = nn.MSELoss()
-
-    '''Initialize datasets and dataloaders:
-        1) Create dataset for later split between training/validation
-        2) Split between training and validation size
-        3) Create train and validation sets and loaders
-    '''
-    train_dataset = GanDataset(
-        config.TRAIN_DIR_IHC, 
-        config.TRAIN_DIR_HE, 
-        config.SUBSET_PERCENTAGE, 
-        patch_size=512, 
-        transform=config.transforms, 
-        shuffle=False
-    )
-    val_dataset = GanDataset(
-        config.TEST_DIR_IHC, 
-        config.TEST_DIR_HE, 
-        config.SUBSET_PERCENTAGE, 
-        patch_size=512, 
-        transform=config.transforms, 
-        shuffle=False,
-    )
-    train_loader = DataLoader(
-        train_dataset, 
-        batch_size=config.BATCH_SIZE, 
-        pin_memory=True, 
-        shuffle=True, 
-        num_workers=config.NUM_WORKERS, 
-        collate_fn=custom_collate,
-        drop_last=True 
-    )
-    val_loader = DataLoader(
-        val_dataset, 
-        batch_size=config.BATCH_SIZE, 
-        pin_memory=True, 
-        shuffle=False, 
-        num_workers=config.NUM_WORKERS, 
-        collate_fn=custom_collate,
-        drop_last=True
-    )
-
-    # Initialize gradient scalers for mixed precision training
-    g_scaler = torch.cuda.amp.GradScaler()
-    d_scaler = torch.cuda.amp.GradScaler()
-
-    start_epoch = 0
-    log_dir = None
-    val_loss = None
-
-    # Load checkpoints if necessary
-    if config.LOAD_MODEL:
-        start_epoch, log_dir, val_loss = load_checkpoint(config.CHECKPOINT_GEN_HE, gen_HE, optim_gen, config.LEARNING_RATE)
-        start_epoch = max(start_epoch, load_checkpoint(config.CHECKPOINT_GEN_IHC, gen_IHC, optim_gen, config.LEARNING_RATE)[0])
-        start_epoch = max(start_epoch, load_checkpoint(config.CHECKPOINT_DISC_HE, disc_HE, optim_disc, config.LEARNING_RATE)[0])
-        start_epoch = max(start_epoch, load_checkpoint(config.CHECKPOINT_DISC_IHC, disc_IHC, optim_disc, config.LEARNING_RATE)[0])
-    
-    if log_dir is None:
-        log_dir = f"logs/GAN_{config.NUM_EPOCHS}_epochs_{datetime.now().strftime('%Y%m%d-%H%M%S')}"
-    if val_loss is None:
-        val_loss = float('inf')
-
-    writer = SummaryWriter(log_dir=log_dir)
-
-    # Early stopping parameters
-    patience = config.EARLY_STOP
-    best_val_loss = val_loss
-    epochs_no_improve = 0
-    best_epoch = 0
-    print(f"Val_loss of loading is {best_val_loss}")
-    # Training loop
-    for epoch in range(start_epoch, config.NUM_EPOCHS):
-        print(f"TRAINING MODEL [Epoch {epoch}]:")
-        gen_train_loss, disc_train_loss = train_func(
-                                            disc_HE, disc_IHC, 
-                                            gen_HE, gen_IHC, 
-                                            optim_disc, optim_gen, 
-                                            g_scaler, d_scaler, 
-                                            cycle_loss, discrim_loss, identity_loss,
-                                            train_loader, 
-                                            epoch, 
-                                            writer
-                                        )
-
-        if epoch % config.FID_FREQUENCY == 0:
-            print(f"CALCULATING FID SCORES [Epoch {epoch}]:")
-            fid_he, fid_ihc = evaluate_fid_scores(gen_HE, gen_IHC, val_loader, config.DEVICE, config.FID_BATCH_SIZE)
-            print(f"\nFID Scores - HE: {fid_he}, IHC: {fid_ihc}\n")
-            writer.add_scalars("FID Scores", {"HE": fid_he, "IHC": fid_ihc}, epoch)
-
-        print(f"VALIDATING MODEL [Epoch {epoch}]:")
-        gen_val_loss = eval_single_epoch(
-                            disc_HE, disc_IHC, 
-                            gen_HE, gen_IHC, 
-                            cycle_loss, discrim_loss, identity_loss,
-                            val_loader, 
-                            epoch, 
-                            writer
-                        )
-
-        writer.add_scalars("Generators Losses", {"train": gen_train_loss, "val": gen_val_loss}, epoch)
-
-        # Check for improvement
-        if gen_val_loss < best_val_loss:
-            best_epoch = epoch
-            best_val_loss = gen_val_loss
-            epochs_no_improve = 0
-            # Save the best model
-            if config.SAVE_MODEL:
-                save_checkpoint(epoch, gen_HE, optim_gen, filename=config.CHECKPOINT_GEN_HE, log_dir=log_dir, loss=best_val_loss)
-                save_checkpoint(epoch, gen_IHC, optim_gen, filename=config.CHECKPOINT_GEN_IHC, log_dir=log_dir, loss=best_val_loss)
-                save_checkpoint(epoch, disc_HE, optim_disc, filename=config.CHECKPOINT_DISC_HE, log_dir=log_dir, loss=best_val_loss)
-                save_checkpoint(epoch, disc_IHC, optim_disc, filename=config.CHECKPOINT_DISC_IHC, log_dir=log_dir, loss=best_val_loss)
-        else:
-            epochs_no_improve += 1
-            print(f"Best epoch so far was {epoch}\n")
-
-        # Check for early stopping
-        if epochs_no_improve >= patience:
-            print("Early stopping triggered")
-            print(f"Last model saved was on epoch {best_epoch} with loss: {best_val_loss}")
-            break
-
-    writer.close()
-
-if __name__ == "__main__":
-    main()
-
-    # # Check if CUDA is available
-    # cuda_available = torch.cuda.is_available()
-
-    # # Print the result
-    # print("CUDA Available:", cuda_available)
-
-    # # If CUDA is available, print the number of GPUs and the current device
-    # if cuda_available:
-    #     num_gpus = torch.cuda.device_count()
-    #     current_device = torch.cuda.current_device()
-    #     device_name = torch.cuda.get_device_name(current_device)
-        
-    #     print("Number of GPUs:", num_gpus)
-    #     print("Current CUDA device:", current_device)
-    #     print("CUDA device name:", device_name)
