--- model/discriminator.py
@@ -1,96 +1,0 @@
-import torch
-import torch.nn as nn
-
-class ConvInstanceNormLeakyReLUBlock(nn.Module):
-    def __init__(self, in_channels: int, out_channels: int, stride: int):
-        """
-        Class object initialization for Convolution-InstanceNorm-LeakyReLU layer
-
-        We use leaky ReLUs with a slope of 0.2.
-        """
-        super().__init__()
-        self.conv = nn.Sequential(
-            nn.Conv2d(
-                in_channels,
-                out_channels,
-                kernel_size=4,
-                stride=stride,
-                padding=1,
-                bias=True,
-                padding_mode="reflect",
-            ),
-            nn.InstanceNorm2d(out_channels),
-            nn.LeakyReLU(0.2, inplace=True),
-        )
-
-    def forward(self, x):
-        return self.conv(x)
-
-
-class Discriminator(nn.Module):
-    def __init__(self, in_channels=3, features=[64, 128, 256, 512]):
-        """
-        Let Ck denote a 4 × 4 Convolution-InstanceNorm-LeakyReLU layer with 
-        k filters and stride 2. Discriminator architecture is: C64-C128-C256-C512. 
-        
-        After the last layer, we apply a convolution to produce a 1-dimensional 
-        output. 
-        
-        We use leaky ReLUs with a slope of 0.2.
-        """
-        super().__init__()
-        self.initial_layer = nn.Sequential(
-            nn.Conv2d(
-                in_channels,
-                features[0],
-                kernel_size=4,
-                stride=2,
-                padding=1,
-                padding_mode="reflect",
-            ),
-            nn.LeakyReLU(0.2, inplace=True),
-        )
-
-        layers = []
-        in_channels = features[0]
-        for feature in features[1:]:
-            layers.append(
-                ConvInstanceNormLeakyReLUBlock(
-                    in_channels, 
-                    feature, 
-                    stride=1 if feature == features[-1] else 2,
-                )
-            )
-            in_channels = feature
-
-        # After the last layer, we apply a convolution to produce a 1-dimensional output 
-        layers.append(
-            nn.Conv2d(
-                in_channels,
-                1,
-                kernel_size=4,
-                stride=1,
-                padding=1,
-                padding_mode="reflect",
-            )
-        )
-        self.model = nn.Sequential(*layers)
-
-    def forward(self, x):
-        x = self.initial_layer(x)
-        # feed the model output into a sigmoid function to make a 1/0 label
-        return torch.sigmoid(self.model(x))
-
-
-def test():  
-    
-    """ Just used to test some features, not applied to training of the model """
-
-    x = torch.randn((5, 3, 512, 512))
-    model = Discriminator(in_channels=3)
-    preds = model(x)
-    print(preds.shape)
-
-
-if __name__ == "__main__":
-    test()
